Args::port is now an unsigned short int
Unsigned short int is >=16 bits in C++. I have changed port form int
to unsigned short int. I have also stop using atoi and begin using
a hardened wrapper to strtol, str2uint16.

A strane behaviour of gcc throws a warning with the following code:

unsigned short int i = 42;
if (uint > 65535) {
    std::cout << i << " is greater than 65535" << std::endl;
} else {
    std::cout << i << " is NOT greater than 65536" << std::endl;
}

It throws a warning saying that i will never be greater than 65536.
Which is true for a 686 system, but can be different for other systems.

TODO: str2uint16 should throws exceptions and go in a separate
header file.

+++ src/args.h
@@ -7,27 +7,27 @@
 #include "llmnr.h"
 
 class Args {
-    unsigned int const d_port;
-    bool         const d_responder;
-    std::string  const d_configFilePath;
-    bool         const d_interactive;
-    bool         const d_debug;
-    std::string  const d_name;
-    std::string  const d_type;
+    unsigned short int  const d_port;
+    bool                const d_responder;
+    std::string         const d_configFilePath;
+    bool                const d_interactive;
+    bool                const d_debug;
+    std::string         const d_name;
+    std::string         const d_type;
     
 public:
 
-    Args(unsigned int const port,
-         bool         const responder,
-         std::string  const &configFilePath,
-         bool         const interactive,
-         bool         const debug,
-         std::string  const &name,
-         std::string  const &type);
+    Args(unsigned short int const port,
+         bool               const responder,
+         std::string        const &configFilePath,
+         bool               const interactive,
+         bool               const debug,
+         std::string        const &name,
+         std::string        const &type);
 
     static Args *       parse(int const argc, char ** const argv);
 
-    int                 port()           const;
+    unsigned short int  port()           const;
     bool                responder()      const;
     std::string const & configFilePath() const;
     bool                interactive()    const;
